+++ CHANGELOG.md
@@ -23,3 +23,17 @@ Added EU PSD2 AISP keywords to package descriptions so people can find it easier
 3. Null Safety implemented.
 
 4. Documentation follows null safety conventions now.
+
+## [1.2.5] - 14th May, 2021
+
+1. Implemented actual full `TransactionData` Model from API Documentation.
+
+2. Renamed `BankAccountModel` to `AccountModel` because of Account schema discrepency in documentation between <https://nordigen.com/en/docs/account-information/overview/parameters-and-responses/> and <https://nordigen.com/en/docs/account-information/output/accounts/>.
+
+3. Implemented `Balance` model from API Documentation.
+
+4. Temporary `getAccountBalancesTemporary` method of `NordigenAccountInfoAPI` class implemented that will be depreciated once the returned data format can be pinned down.
+
+5. Lint -> Pedantic.
+
+6. TODO: Resolve account model discrepency between <https://nordigen.com/en/docs/account-information/output/accounts/> and Schema given in <https://nordigen.com/en/docs/account-information/overview/parameters-and-responses/>